Add initial PHPStan config (#212)

+++ src/Check/Autoload_Options_Size.php
@@ -17,6 +17,9 @@ class Autoload_Options_Size extends Check {
 	 */
 	protected $threshold_kb = 900;
 
+	/**
+	 * @return void
+	 */
 	public function run() {
 		ob_start();
 		WP_CLI::run_command(
@@ -40,9 +43,14 @@ class Autoload_Options_Size extends Check {
 		}
 	}
 
+	/**
+	 * @param int|float $size Size in bytes.
+	 * @param int       $precision Precision.
+	 * @return string
+	 */
 	private static function format_bytes( $size, $precision = 2 ) {
 		$base     = log( $size, 1024 );
 		$suffixes = array( '', 'kb', 'mb', 'g', 't' );
-		return round( pow( 1024, $base - floor( $base ) ), $precision ) . $suffixes[ floor( $base ) ];
+		return round( pow( 1024, $base - floor( $base ) ), $precision ) . $suffixes[ (int) floor( $base ) ];
 	}
 }
